Move functions around into badger_restore
