+++ backupstore.go
@@ -21,14 +21,11 @@ import (
 	"bytes"
 	"encoding/binary"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"os"
 	"path/filepath"
-	"sort"
 
 	"github.com/dgraph-io/badger/protos"
-	"github.com/dgraph-io/badger/y"
 	"github.com/pkg/errors"
 )
 
@@ -75,174 +72,8 @@ func ReadBackupStatus(path string) (protos.BackupStatus, error) {
 	return ret, nil
 }
 
-func backupFileName(id uint64) string {
+func BackupFileName(id uint64) string {
 	return fmt.Sprintf("backup-%06d", id)
-}
-
-// Implements the y.Iterator interface.
-type backupFileIterator struct {
-	fp     *os.File
-	reader *bufio.Reader
-	key    []byte
-	value  y.ValueStruct
-	err    error
-}
-
-func makeBackupFileIterator(path string, id uint64, bufferSize int) (*backupFileIterator, error) {
-	fp, err := os.Open(filepath.Join(path, backupFileName(id)))
-	if err != nil {
-		return nil, err
-	}
-	return &backupFileIterator{fp: fp}, nil
-}
-
-func (bit *backupFileIterator) Rewind() {
-	if bit.err != nil {
-		bit.err = nil
-	}
-	_, err := bit.fp.Seek(0, os.SEEK_SET)
-	if err != nil {
-		bit.err = err
-		bit.key = nil
-		bit.value = y.ValueStruct{}
-		return
-	}
-	bit.reader = bufio.NewReader(bit.fp)
-	bit.Next()
-}
-
-func (bit *backupFileIterator) Key() []byte {
-	return bit.key
-}
-
-func (bit *backupFileIterator) Value() y.ValueStruct {
-	return bit.value
-}
-
-func (bit *backupFileIterator) Next() {
-	if bit.err != nil {
-		return
-	}
-	sz, err := binary.ReadUvarint(bit.reader)
-	if err != nil {
-		bit.err = err
-		bit.key = nil
-		bit.value = y.ValueStruct{}
-		return
-	}
-
-	data := make([]byte, sz)
-	if _, err := io.ReadFull(bit.reader, data); err != nil {
-		if err == io.EOF {
-			err = io.ErrUnexpectedEOF
-		}
-		bit.err = err
-		bit.key = nil
-		bit.value = y.ValueStruct{}
-		return
-	}
-
-	var item protos.BackupItem
-	if err := item.Unmarshal(data); err != nil {
-		bit.err = err
-		bit.key = nil
-		bit.value = y.ValueStruct{}
-		return
-	}
-	bit.key = item.Key
-	var meta byte
-	if !item.HasValue {
-		meta = BitDelete
-	}
-	bit.value = y.ValueStruct{
-		// This is typically a vptr but in our case it's a value -- we're just using MergeIterator.
-		Value:      item.Value,
-		Meta:       meta,
-		UserMeta:   uint8(item.UserMeta),
-		CASCounter: item.Counter,
-	}
-}
-
-func (bit *backupFileIterator) Valid() bool {
-	return bit.key != nil
-}
-
-func (bit *backupFileIterator) Close() error {
-	err := bit.err
-	if err == io.EOF {
-		err = nil
-	}
-	if closeErr := bit.fp.Close(); err == nil {
-		err = closeErr
-	}
-	return err
-}
-
-func (bit *backupFileIterator) Seek([]byte) {
-	panic("Seek not implemented")
-}
-
-// RetrieveBackup streams all changes from the backup store in increasing key order to consumer.
-func RetrieveBackup(path string, consumer func(protos.BackupItem) error) (err error) {
-	status, err := ReadBackupStatus(path)
-	if err != nil {
-		return err
-	}
-
-	ids := []uint64{}
-	for _, backup := range status.Backups {
-		ids = append(ids, backup.BackupID)
-	}
-
-	// Sort by descending ID (because MergeIterator expects order by increasing age)
-	sort.Slice(ids, func(i, j int) bool { return ids[i] > ids[j] })
-
-	const memUsage = 2 << 30
-	bufferSize := memUsage / len(ids)
-
-	fileIters := []y.Iterator{}
-	defer func() {
-		for _, iter := range fileIters {
-			_ = iter.Close()
-		}
-	}()
-
-	for _, id := range ids {
-		iter, err := makeBackupFileIterator(path, id, bufferSize)
-		if err != nil {
-			return err
-		}
-		fileIters = append(fileIters, iter)
-	}
-
-	// In MergeIterator, keys returned by the iterators closer to the beginning of the array take
-	// precedence.
-	mergeIter := y.NewMergeIterator(fileIters, false)
-	fileIters = nil
-	defer func() {
-		if closeErr := mergeIter.Close(); err == nil {
-			err = closeErr
-		}
-	}()
-
-	for mergeIter.Rewind(); mergeIter.Valid(); mergeIter.Next() {
-		key := mergeIter.Key()
-		value := mergeIter.Value()
-		if value.Meta == 0 { // if value.Meta != BitDelete
-			err := consumer(protos.BackupItem{
-				Key:      key,
-				Counter:  value.CASCounter,
-				HasValue: true,
-				UserMeta: uint32(value.UserMeta),
-				Value:    value.Value,
-			})
-			if err != nil {
-				return err
-			}
-		}
-	}
-
-	return nil
 }
 
 // NewBackup stores a new backup onto an existing backup store.  producer is deemed finished when
@@ -323,7 +154,7 @@ func NewBackup(path string, counter uint64,
 	}
 
 	newBackupID := maxBackupID + 1
-	err = syncCloseRename(f, f.Name(), path, backupFileName(newBackupID))
+	err = syncCloseRename(f, f.Name(), path, BackupFileName(newBackupID))
 	f = nil
 	if err != nil {
 		return err
